+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Change log
 `typerep-map` uses [PVP Versioning][1].
 The change log is available [on GitHub][2].
 
+# 0.2.0
+
+* [#43](https://github.com/kowainik/typerep-map/issues/43):
+  Implement `IsList` instance for `TypeRepMap`.
+  Add `hoistA` function.
+* Drop support for `ghc-8.0.2`.
+
 # 0.1.0
 
 * Initially created.